Added db owner to all scripts... Please look through an VERIFY that all of them have it... I'm sure I missed a few.
Modified installation to support the replacement of {databaseOwner} throughout the script.

+++ yafsrc/YetAnotherForum.NET/install/views.sql
@@ -2,25 +2,25 @@
 ** Views
 */
 
-if exists (select 1 from sysobjects where id = object_id(N'yaf_vaccess') and OBJECTPROPERTY(id, N'IsView') = 1)
-	drop view yaf_vaccess
+if exists (select 1 from sysobjects where id = object_id(N'[{databaseOwner}].[yaf_vaccess]') and OBJECTPROPERTY(id, N'IsView') = 1)
+	drop view [{databaseOwner}].[yaf_vaccess]
 GO
 
-if exists (select 1 from sysobjects where id = object_id(N'yaf_vmaccess') and OBJECTPROPERTY(id, N'IsView') = 1)
-	drop view yaf_vmaccess
+if exists (select 1 from sysobjects where id = object_id(N'[{databaseOwner}].[yaf_vmaccess]') and OBJECTPROPERTY(id, N'IsView') = 1)
+	drop view [{databaseOwner}].[yaf_vmaccess]
 GO
 
-if exists (select 1 from sysobjects where id = object_id(N'yaf_PMessageView') and OBJECTPROPERTY(id, N'IsView') = 1)
-	drop view yaf_PMessageView
+if exists (select 1 from sysobjects where id = object_id(N'[{databaseOwner}].[yaf_PMessageView]') and OBJECTPROPERTY(id, N'IsView') = 1)
+	drop view [{databaseOwner}].[yaf_PMessageView]
 GO
 
-create view dbo.yaf_vaccess as
+create view [{databaseOwner}].[yaf_vaccess] as
 	select
 		UserID				= a.UserID,
 		ForumID				= x.ForumID,
 		IsAdmin				= max(convert(int,b.Flags & 1)),
 		IsForumModerator	= max(convert(int,b.Flags & 8)),
-		IsModerator			= (select count(1) from dbo.yaf_UserGroup v,dbo.yaf_Group w,dbo.yaf_ForumAccess x,dbo.yaf_AccessMask y where v.UserID=a.UserID and w.GroupID=v.GroupID and x.GroupID=w.GroupID and y.AccessMaskID=x.AccessMaskID and (y.Flags & 64)<>0),
+		IsModerator			= (select count(1) from [{databaseOwner}].[yaf_UserGroup] v,[{databaseOwner}].[yaf_Group] w,[{databaseOwner}].[yaf_ForumAccess] x,[{databaseOwner}].[yaf_AccessMask] y where v.UserID=a.UserID and w.GroupID=v.GroupID and x.GroupID=w.GroupID and y.AccessMaskID=x.AccessMaskID and (y.Flags & 64)<>0),
 		ReadAccess			= max(x.ReadAccess),
 		PostAccess			= max(x.PostAccess),
 		ReplyAccess			= max(x.ReplyAccess),
@@ -46,8 +46,8 @@ create view dbo.yaf_vaccess as
 			DeleteAccess		= convert(int,c.Flags & 256),
 			UploadAccess		= convert(int,c.Flags & 512)
 		from
-			dbo.yaf_UserForum b
-			join dbo.yaf_AccessMask c on c.AccessMaskID=b.AccessMaskID
+			[{databaseOwner}].[yaf_UserForum] b
+			join [{databaseOwner}].[yaf_AccessMask] c on c.AccessMaskID=b.AccessMaskID
 		
 		union
 		
@@ -65,9 +65,9 @@ create view dbo.yaf_vaccess as
 			DeleteAccess		= convert(int,d.Flags & 256),
 			UploadAccess		= convert(int,d.Flags & 512)
 		from
-			dbo.yaf_UserGroup b
-			join dbo.yaf_ForumAccess c on c.GroupID=b.GroupID
-			join dbo.yaf_AccessMask d on d.AccessMaskID=c.AccessMaskID
+			[{databaseOwner}].[yaf_UserGroup] b
+			join [{databaseOwner}].[yaf_ForumAccess] c on c.GroupID=b.GroupID
+			join [{databaseOwner}].[yaf_AccessMask] d on d.AccessMaskID=c.AccessMaskID
 
 		union
 
@@ -85,19 +85,19 @@ create view dbo.yaf_vaccess as
 			DeleteAccess	= convert(int,0),
 			UploadAccess	= convert(int,0)
 		from
-			dbo.yaf_User a
+			[{databaseOwner}].[yaf_User] a
 		) as x
-		join dbo.yaf_UserGroup a on a.UserID=x.UserID
-		join dbo.yaf_Group b on b.GroupID=a.GroupID
+		join [{databaseOwner}].[yaf_UserGroup] a on a.UserID=x.UserID
+		join [{databaseOwner}].[yaf_Group] b on b.GroupID=a.GroupID
 	group by a.UserID,x.ForumID
 GO
 
-create view [dbo].[yaf_vmaccess] as
+create view [{databaseOwner}].[yaf_vmaccess] as
 	select
 		UserID				= a.UserID,
 		ForumID				= x.ForumID,
 		IsForumModerator	= max(convert(int,b.Flags & 8)),
-		IsModerator			= (select count(1) from dbo.yaf_UserGroup v,dbo.yaf_Group w,dbo.yaf_ForumAccess x,dbo.yaf_AccessMask y where v.UserID=a.UserID and w.GroupID=v.GroupID and x.GroupID=w.GroupID and y.AccessMaskID=x.AccessMaskID and (y.Flags & 64)<>0),
+		IsModerator			= (select count(1) from [{databaseOwner}].[yaf_UserGroup] v,[{databaseOwner}].[yaf_Group] w,[{databaseOwner}].[yaf_ForumAccess] x,[{databaseOwner}].[yaf_AccessMask] y where v.UserID=a.UserID and w.GroupID=v.GroupID and x.GroupID=w.GroupID and y.AccessMaskID=x.AccessMaskID and (y.Flags & 64)<>0),
 		ReadAccess			= max(x.ReadAccess),
 		PostAccess			= max(x.PostAccess),
 		ReplyAccess			= max(x.ReplyAccess),
@@ -123,8 +123,8 @@ create view [dbo].[yaf_vmaccess] as
 			DeleteAccess		= convert(int,c.Flags & 256),
 			UploadAccess		= convert(int,c.Flags & 512)
 		from
-			dbo.yaf_UserForum b
-			join dbo.yaf_AccessMask c on c.AccessMaskID=b.AccessMaskID
+			[{databaseOwner}].[yaf_UserForum] b
+			join [{databaseOwner}].[yaf_AccessMask] c on c.AccessMaskID=b.AccessMaskID
 		
 		union
 		
@@ -142,10 +142,10 @@ create view [dbo].[yaf_vmaccess] as
 			DeleteAccess		= convert(int,d.Flags & 256),
 			UploadAccess		= convert(int,d.Flags & 512)
 		from
-			dbo.yaf_UserGroup b
-			join dbo.yaf_ForumAccess c on c.GroupID=b.GroupID
-			join dbo.yaf_AccessMask d on d.AccessMaskID=c.AccessMaskID
-			join dbo.yaf_Group e on e.GroupID=b.GroupID
+			[{databaseOwner}].[yaf_UserGroup] b
+			join [{databaseOwner}].[yaf_ForumAccess] c on c.GroupID=b.GroupID
+			join [{databaseOwner}].[yaf_AccessMask] d on d.AccessMaskID=c.AccessMaskID
+			join [{databaseOwner}].[yaf_Group] e on e.GroupID=b.GroupID
 		where 
 			(e.Flags & 1) = 0 --is not admin group
 
@@ -165,24 +165,24 @@ create view [dbo].[yaf_vmaccess] as
 			DeleteAccess	= convert(int,0),
 			UploadAccess	= convert(int,0)
 		from
-			dbo.yaf_User a
+			[{databaseOwner}].[yaf_User] a
 		) as x
-		join dbo.yaf_UserGroup a on a.UserID=x.UserID
-		join dbo.yaf_Group b on b.GroupID=a.GroupID
+		join [{databaseOwner}].[yaf_UserGroup] a on a.UserID=x.UserID
+		join [{databaseOwner}].[yaf_Group] b on b.GroupID=a.GroupID
 
 	group by a.UserID,x.ForumID
 GO
 
 -- yaf_PMessageView
 
-CREATE VIEW [dbo].[yaf_PMessageView]
+CREATE VIEW [{databaseOwner}].[yaf_PMessageView]
 AS
-SELECT     dbo.yaf_PMessage.PMessageID, dbo.yaf_UserPMessage.UserPMessageID, dbo.yaf_PMessage.FromUserID, yaf_User_From.Name AS FromUser, 
-                      dbo.yaf_UserPMessage.UserID AS ToUserId, dbo.yaf_User.Name AS ToUser, dbo.yaf_PMessage.Created, dbo.yaf_PMessage.Subject, 
-                      dbo.yaf_PMessage.Body, dbo.yaf_PMessage.Flags, dbo.yaf_UserPMessage.IsRead, dbo.yaf_UserPMessage.IsInOutbox, dbo.yaf_UserPMessage.IsArchived
-FROM         dbo.yaf_PMessage INNER JOIN
-                      dbo.yaf_UserPMessage ON dbo.yaf_PMessage.PMessageID = dbo.yaf_UserPMessage.PMessageID INNER JOIN
-                      dbo.yaf_User ON dbo.yaf_UserPMessage.UserID = dbo.yaf_User.UserID INNER JOIN
-                      dbo.yaf_User AS yaf_User_From ON dbo.yaf_PMessage.FromUserID = yaf_User_From.UserID
+SELECT     [{databaseOwner}].[yaf_PMessage].PMessageID, [{databaseOwner}].[yaf_UserPMessage].UserPMessageID, [{databaseOwner}].[yaf_PMessage].FromUserID, yaf_User_From.Name AS FromUser, 
+                      [{databaseOwner}].[yaf_UserPMessage].UserID AS ToUserId, [{databaseOwner}].[yaf_User].Name AS ToUser, [{databaseOwner}].[yaf_PMessage].Created, [{databaseOwner}].[yaf_PMessage].Subject, 
+                      [{databaseOwner}].[yaf_PMessage].Body, [{databaseOwner}].[yaf_PMessage].Flags, [{databaseOwner}].[yaf_UserPMessage].IsRead, [{databaseOwner}].[yaf_UserPMessage].IsInOutbox, [{databaseOwner}].[yaf_UserPMessage].IsArchived
+FROM         [{databaseOwner}].[yaf_PMessage] INNER JOIN
+                      [{databaseOwner}].[yaf_UserPMessage] ON [{databaseOwner}].[yaf_PMessage].PMessageID = [{databaseOwner}].[yaf_UserPMessage].PMessageID INNER JOIN
+                      [{databaseOwner}].[yaf_User] ON [{databaseOwner}].[yaf_UserPMessage].UserID = [{databaseOwner}].[yaf_User].UserID INNER JOIN
+                      [{databaseOwner}].[yaf_User] AS yaf_User_From ON [{databaseOwner}].[yaf_PMessage].FromUserID = yaf_User_From.UserID
 
 GO
